FP8 CRITICAL FIX: Convert scales to float32 for AITER bpreshuffle kernel
AITER requires float32 scales but was receiving bfloat16, causing failures.
Converted both x_scale and w_scale to float32 before calling AITER GEMM.

+++ python/sglang/srt/layers/quantization/fp8_utils.py
@@ -606,8 +606,9 @@ def apply_fp8_linear(
             )
             
             # Reshape scales for AITER (expects [M,1] and [N,1])
-            x_scale_aiter = x_scale.view(-1, 1)
-            w_scale_aiter = weight_scale.view(-1, 1)  # [N, 1] not [1, N]!
+            # FP8 CRITICAL FIX: AITER kernel requires FLOAT32 scales!
+            x_scale_aiter = x_scale.view(-1, 1).to(torch.float32)
+            w_scale_aiter = weight_scale.view(-1, 1).to(torch.float32)  # [N, 1] not [1, N]!
             
             # DEBUG: Log first call details
             if not hasattr(apply_fp8_linear, '_logged_once'):
